test: 修改之前 skip 掉的测试用例
