+++ packages/f2/test/components/point/point.test.tsx
@@ -7,7 +7,7 @@ import { createContext } from '../../util';
 const url = 'https://gw.alipayobjects.com/os/antfincdn/6HodecuhvM/scatter.json';
 const url2 = 'https://gw.alipayobjects.com/os/antfincdn/aN68ysvGFa/index.json';
 
-describe.skip('Point Chart', () => {
+describe('Point Chart', () => {
   it('基础点图', async () => {
     const res = await fetch(url);
     const data = await res.json();
@@ -54,8 +54,7 @@ describe.skip('Point Chart', () => {
       </Canvas>
     );
 
-    // @ts-ignore
-    const canvas = new type(props);
+    const canvas = new Canvas(props);
     canvas.render();
   });
 
@@ -87,8 +86,7 @@ describe.skip('Point Chart', () => {
       </Canvas>
     );
 
-    // @ts-ignore
-    const canvas = new type(props);
+    const canvas = new Canvas(props);
     canvas.render();
   });
 });
